refactored to use QR code generator at frontend, and styled the component
